Add book forms and adjust pricing
added Django ModelForms for books, authors, publishers, categories, and a search/filter form for listing books updated the Book.price field from DecimalField to PositiveIntegerField to match the app's integer-based pricing model.

+++ books/models.py
@@ -44,7 +44,7 @@ class Book(models.Model):
     publisher = models.ForeignKey(Publisher, on_delete=models.PROTECT)
     description = models.TextField()
     language = models.CharField(max_length=5, choices=LANGUAGE_CHOICES)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.PositiveIntegerField()
     pages = models.PositiveSmallIntegerField()
     isbn = models.CharField(max_length=20, blank=True, unique=True)
     publish_date = models.DateField()
